# lib/api-zod/src/generated/types/couple.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Our Hearts couples app API
 * OpenAPI spec version: 0.1.0
 */
import type { PartnerInfo } from './partnerInfo';

export interface Couple {
  id: number;
  /** ISO date string YYYY-MM-DD */
  startDate: string;
  me: PartnerInfo;
  partner: PartnerInfo;
}