# lib/api-zod/src/generated/types/dailyNote.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Our Hearts couples app API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyNote {
  id: number;
  userId: number;
  coupleId: number;
  displayName: string;
  avatarColor: string;
  content: string;
  date: string;
  createdAt: string;
}